Added documentation to the new features
