+++ spring-cloud-contract-verifier/src/test/groovy/org/springframework/cloud/contract/verifier/GeneratorScannerSpec.groovy
@@ -16,14 +16,14 @@
 
 package org.springframework.cloud.contract.verifier
 
-import org.springframework.cloud.contract.verifier.builder.SingleTestGenerator
+import org.springframework.cloud.contract.verifier.builder.JavaTestGenerator
 import org.springframework.cloud.contract.verifier.config.ContractVerifierConfigProperties
 import org.springframework.cloud.contract.verifier.config.TestFramework
 import spock.lang.Specification
 
 class GeneratorScannerSpec extends Specification {
 
-	private SingleTestGenerator classGenerator = Mock(SingleTestGenerator)
+	private JavaTestGenerator classGenerator = Mock(JavaTestGenerator)
 
 	def "should find all .json files and generate 6 classes for them"() {
 		given:
@@ -34,7 +34,7 @@ class GeneratorScannerSpec extends Specification {
 		when:
 			testGenerator.generateTestClasses("org.springframework.cloud.contract.verifier")
 		then:
-			6 * classGenerator.buildClass(_, _, _, _) >> "qwerty"
+			6 * classGenerator.buildClass(_, _, _, _, _) >> "qwerty"
 	}
 
 	def "should create class with full package"() {
@@ -45,9 +45,9 @@ class GeneratorScannerSpec extends Specification {
 		when:
 			testGenerator.generateTestClasses("org.springframework.cloud.contract.verifier")
 		then:
-			1 * classGenerator.buildClass(_, 'exceptionsSpec', 'org.springframework.cloud.contract.verifier', _) >> "spec"
-			1 * classGenerator.buildClass(_, 'exceptionsSpec', 'org.springframework.cloud.contract.verifier.v1', _) >> "spec1"
-			1 * classGenerator.buildClass(_, 'exceptionsSpec', 'org.springframework.cloud.contract.verifier.v2', _) >> "spec2"
+			1 * classGenerator.buildClass(_, _, 'exceptionsSpec', 'org.springframework.cloud.contract.verifier', _) >> "spec"
+			1 * classGenerator.buildClass(_, _, 'exceptionsSpec', 'org.springframework.cloud.contract.verifier.v1', _) >> "spec1"
+			1 * classGenerator.buildClass(_, _, 'exceptionsSpec', 'org.springframework.cloud.contract.verifier.v2', _) >> "spec2"
 	}
 
 }
